refactor: 패키지 구조 변경 (#2)
- 도메인형 패키지 구조로 리팩토링함
- 덩치가 너무 커진 StargazingService 클래스에 Facade 패턴 적용할 예정
