+++ src/main/java/xyz/ncookie/stargazer/domain/stargazing/controller/StargazingController.java
@@ -1,4 +1,4 @@
-package xyz.ncookie.stargazer.controller;
+package xyz.ncookie.stargazer.domain.stargazing.controller;
 
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -7,10 +7,10 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import lombok.RequiredArgsConstructor;
-import xyz.ncookie.stargazer.dto.StargazingForecastResponse;
-import xyz.ncookie.stargazer.dto.StargazingRequest;
-import xyz.ncookie.stargazer.dto.StargazingResponse;
-import xyz.ncookie.stargazer.service.StargazingService;
+import xyz.ncookie.stargazer.domain.stargazing.dto.StargazingForecastResponse;
+import xyz.ncookie.stargazer.domain.stargazing.dto.StargazingRequest;
+import xyz.ncookie.stargazer.domain.stargazing.dto.StargazingResponse;
+import xyz.ncookie.stargazer.domain.stargazing.service.StargazingService;
 
 @RestController
 @RequestMapping("/api/v1")
