+++ src/main/java/xyz/ncookie/stargazer/domain/stargazing/dto/StargazingResponse.java
@@ -1,4 +1,4 @@
-package xyz.ncookie.stargazer.dto;
+package xyz.ncookie.stargazer.domain.stargazing.dto;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
